% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianmasem-package.R
\docType{package}
\name{bayesianmasem-package}
\alias{bayesianmasem-package}
\alias{bayesianmasem}
\title{The 'bayesianmasem' package.}
\description{
This is a package for Bayesian meta-analytic
structural equation models (MASEMs), specifically confirmatory factor models.
The package fits fixed-, random- and dependent-samples MASEMs.
}
