% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmasem_ci_results.R
\name{bmasem_ci_results}
\alias{bmasem_ci_results}
\title{Conditional indepepence estimates for path analysis}
\usage{
bmasem_ci_results(object, interval = 0.9, summarize = TRUE)
}
\arguments{
\item{object}{(bmasem) An object of \code{\link{bmasem-class}}
returned by \code{\link{bayesianmasem}}.}

\item{interval}{Confidence interval to select}

\item{summarize}{(LOGICAL)
If TRUE (default): Return posterior summary as data.frame;
If FALSE: Return posterior draws as data.frame.}
}
\value{
Table of parameters related to conditional independence.
}
\description{
Obtain conditional independence estimates for path analysis.
Estimates are standardized.
}
\examples{
\dontrun{
model_syntax <- paste(
  "BI ~ ATT + SN + PBC",
  "BEH ~ PBC + BI",
  sep = "\n"
)
fit <- bmasem(
  model_syntax,
  sample_cov = cooke16$data, sample_nobs = cooke16$n
)
ci_table <- bmasem_ci_results(fit)
}
}
