% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmasem_composite_reliability.R
\name{bmasem_composite_reliability}
\alias{bmasem_composite_reliability}
\title{Obtain composite reliability from a fitted model}
\usage{
bmasem_composite_reliability(object, interval = 0.9, return_draws = FALSE)
}
\arguments{
\item{object}{(bmasem) An object of \code{\link{bmasem-class}}
returned by \code{\link{bmasem}}.}

\item{interval}{(real in (0, 1)) Credible interval to return.}

\item{return_draws}{(LOGICAL) If \code{TRUE}, returns the realised distribution
of reliability.
If \code{FALSE} (default), return the distribution summary.}
}
\value{
\itemize{
\item If \code{return_draws = FALSE},
returns the summary of posterior distribution of composite reliability
for each factor.
\item If \code{return_draws = TRUE},
returns draws from the realised distribution of composite reliability
for each factor.
}
}
\description{
Get composite reliability
}
\details{
This metric is the ratio of common variance (CV) to
total variance (TV) for each factor.
\itemize{
\item \strong{Common Variance}: For factor \( f \), the common variance is:
[
\mathrm{CV}\emph{f =
\left(\sum}{i \in I_f} \lambda_{if}\right)^2 \phi_{ff},
]
where \( I_f \) indexes the indicators of factor \( f \).
\item \strong{Total Variance}: For factor \( f \), the total variance is:
[
\mathrm{TV}\emph{f =
\sum}{i \in I_f} \sum_{j \in I_f} \omega_{ij},
]
where \( \Omega = = (\omega_{ij}) \) is the model-implied covariance
matrix of the observed indicators.
}
}
\examples{
\dontrun{
model_syntax <- paste0(
  "distress =~ ", paste0("x", 1:14, collapse = " + "), "\n",
  "anxiety =~ ", paste0("x", seq(1, 14, 2), collapse = " + "), "\n",
  "depression =~ ", paste0("x", seq(2, 14, 2), collapse = " + ")
)
fit <- bmasem(
  model_syntax,
  sample_cov = Norton13$data, sample_nobs = Norton13$n, orthogonal = TRUE
)
bmasem_composite_reliability(fit)
}
}
