% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmasem_plot.R
\name{bmasem_plot}
\alias{bmasem_plot}
\title{Pretty print model results}
\usage{
bmasem_plot(object, type = "trace", subset = NULL, ...)
}
\arguments{
\item{object}{(bmasem) An object of \code{\link{bmasem-class}}
returned by \code{\link{bmasem}}.}

\item{type}{(character) Type of plot:
"trace" for traceplots and "hist" for histograms.}

\item{subset}{(character) Subset of parameters:
NULL (Default) showing all estimated parameters;
Any other response will be used as regular expressions to subset
the parameters. It can be loading names or types of parameters.}

\item{...}{additional arguments to relevant bayesplot function}
}
\value{
bayesplot object
}
\description{
Plot model parameters
}
\examples{
\dontrun{
model_syntax <- paste0(
  "distress =~ ", paste0("x", 1:14, collapse = " + "), "\n",
  "anxiety =~ ", paste0("x", seq(1, 14, 2), collapse = " + "), "\n",
  "depression =~ ", paste0("x", seq(2, 14, 2), collapse = " + ")
)
bmasem(
  model_syntax,
  sample_cov = Norton13$data, sample_nobs = Norton13$n, orthogonal = TRUE
)
bmasem_plot(fit)
}
}
