% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmasem_residual_network.R
\name{bmasem_residual_network}
\alias{bmasem_residual_network}
\title{Compute residual network}
\usage{
bmasem_residual_network(object)
}
\arguments{
\item{object}{(bmasem) An object of \code{\link{bmasem-class}}
returned by \code{\link{bmasem}}.}
}
\value{
A data.frame containing posterior samples of the partial
correlation matrix.
}
\description{
Interpret the error correlations as a residual network model
\insertCite{epskamp_generalized_2017}{bayesianmasem}.
}
\examples{
\dontrun{
model_syntax <- paste("distress =~", paste0("x", 1:14, collapse = " + "))
fit <- bmasem(
  model_syntax,
  sample_cov = Norton13$data, sample_nobs = Norton13$n
)
res_net <- bmasem_residual_network(fit)
p_corr_df <- posterior::summarise_draws(res_net)
n_items <- sqrt(nrow(p_corr_df))
p_corr_mat <- matrix(p_corr_df$mean, n_items)
p_corr_mat
qgraph::qgraph(p_corr_mat)
}
}
