% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmasem_stage_2.R
\name{bmasem_stage_2}
\alias{bmasem_stage_2}
\title{Fit random-effects Bayesian meta-analytic CFAs with minor factors assumed.}
\usage{
bmasem_stage_2(
  model = NULL,
  pool_fit = NULL,
  method = "normal",
  orthogonal = FALSE,
  simple_struc = TRUE,
  seed = 12345,
  warmup = 1000,
  sampling = 1000,
  refresh = (warmup + sampling)/10,
  adapt_delta = 0.9,
  max_treedepth = 10,
  chains = 3,
  ncores = max(parallel::detectCores() - 2, 1),
  priors = new_bmasempriors(),
  show = TRUE,
  show_messages = TRUE,
  cluster = NULL
)
}
\arguments{
\item{model}{A description of the user-specified model, lavaan syntax.}

\item{pool_fit}{The list returned by the \code{bmasem_stage_1} function.}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", or "none". See details below.}

\item{orthogonal}{(LOGICAL)
If TRUE: constrain all factors orthogonal (overrides model syntax);
If FALSE (default): according to model syntax.}

\item{simple_struc}{(LOGICAL) Only relevant for CFAs.
If TRUE (default): assume simple structure;
If FALSE: estimate all cross-loadings using generalized}

\item{seed}{(positive integer) seed, set to obtain replicable results.}

\item{warmup}{(positive integer) The number of warmup iterations to run per
chain.}

\item{sampling}{(positive integer) The number of post-warmup iterations to
run per chain, retained for inference.}

\item{refresh}{(positive integer) How often to print the status
of the sampler.}

\item{adapt_delta}{(real in (0, 1)) Increase to resolve divergent
transitions.}

\item{max_treedepth}{(positive integer) Increase to resolve problems with
maximum tree depth.}

\item{chains}{(positive integer) The number of Markov chains to run.}

\item{ncores}{(positive integer) The number of chains to run in parallel.}

\item{priors}{An object of \code{\link{bmasempriors-class}}.
See \code{\link{new_bmasempriors}} for more information.}

\item{show}{(Logical) If TRUE, show table of results, if FALSE, do not
show table of results. As an example, use FALSE for simulation studies.}

\item{show_messages}{(Logical) If TRUE, show messages from Stan sampler,
if FALSE, hide messages.}

\item{cluster}{An optional integer vector identifying the cluster each group
belongs to.
Asssume there are five groups, the first three belong to cluster 1
and the last two belong to cluster 2,
then the argument would be: \code{cluster = c(1, 1, 1, 2, 2)}.
This feature is experimental, see details below.}
}
\value{
An object of \code{\link{bmasem-class}}
}
\description{
A function that takes a pooled correlation matrix object
and returns a fitted SEM.
}
\details{
CFAs assume standardized factors.
Latent variable regression models are not yet implemented.

There are different methods for estimating models in this package:
\itemize{
\item \code{normal}: under belief that minor factor influences are
on average zero with continuous deviations away from zero.
\item \code{lasso}: under belief that minor factor influences are largely
zero with a small number of non-zero residual covariances.
\item \code{logistic}: for similar belief as normal but more readily
accomodates extreme outliers.
\item \code{GDP}: to mimic a global-local approach, i.e.
attempt to shrink near 0 residual covariances to 0
with minimal shrinking for larger residual covariances
\insertCite{armagan_generalized_2013}{bayesianmasem}.
\item \code{none}: if intending to ignore the influence of minor factors.
}
}
\examples{
\dontrun{
pool_fit_issp <- bmasem_stage_1(
  sample_cov = issp89$data, sample_nobs = issp89$n
)
model_syntax_issp <- "# latent variable definitions
F1 =~ JP1 + JP2 + JP3
F2 =~ JN1 + JN2 + JN4 + JN4
F3 =~ TD1 + TD2"
bmasem_stage_2(model_syntax_issp, pool_fit_issp)
pool_fit_norton <- bmasem_stage_1(
  sample_cov = Norton13$data, sample_nobs = Norton13$n
)
model_syntax_norton <- paste0(
  "distress =~ ", paste0("x", 1:14, collapse = " + "), "\n",
  "anxiety =~ ", paste0("x", seq(1, 14, 2), collapse = " + "), "\n",
  "depression =~ ", paste0("x", seq(2, 14, 2), collapse = " + ")
)
bmasem_stage_2(model_syntax_norton, pool_fit_norton, orthogonal = TRUE)
}
}
\references{
\insertAllCited{}
}
