% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_defs.R
\docType{class}
\name{bmasempriors-class}
\alias{bmasempriors-class}
\title{A class for setting up priors.}
\description{
A class for setting up priors.
}
\section{Slots}{

\describe{
\item{\code{lkj_shape}}{(positive real) The shape parameter of the LKJ-prior on the
interfactor correlation matrix in confirmatory factor models.}

\item{\code{ml_par}}{(real) The location parameter of the normal prior on loadings.}

\item{\code{sl_par}}{(positive real) The scale parameter
of the normal prior on loadings.}

\item{\code{rs_par}}{(positive real) The scale parameter
of the Student-t(3,0,) prior on residual standard deviations.}

\item{\code{rc_par}}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
prior on the residual error correlations.}

\item{\code{mr_par}}{(real) The location parameter of the normal prior
on the log-RMSEA.}

\item{\code{sr_par}}{(positive real) The scale parameter of the normal prior
on the log-RMSEA.}

\item{\code{br_par}}{(positive real) The scale parameter of the normal prior
on the regression coefficients for the log-RMSEA.}

\item{\code{rm_par}}{(positive real) The scale parameter of the normal prior
on the tau / CRMR parameter.}
}}

