% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boer16}
\alias{boer16}
\title{Another dataset from the metaSEM package.}
\format{
\subsection{\code{boer16}}{

A list of data with the following structure:
\describe{
\item{data}{
A list of correlation matrices. The variables are LMX
(leader-member exchange), TFL (transformational leadership),
JS (job satisfaction), OC (organizational commitment), and
LE (leader effectiveness).
}
\item{n}{A vector of sample sizes}
\item{RelLMX}{The reliability of LMX}
\item{RelTFL}{The reliability of TFL}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/metaSEM/index.html}
}
\usage{
boer16
}
\description{
The data set includes 131 correlation matrices on the
mediating role of leader–member exchange in transformational
leadership
\insertCite{boer_revisiting_2016}{bayesianmasem}.
Data and documentation copied over from the metaSEM package
\insertCite{Cheung-metaSEM}{bayesianmasem}.
}
\examples{
\dontrun{
model_syntax <- paste(
  "LMX ~ TFL",
  "JS + OC + LE ~ TFL + LMX",
  sep = "\n"
)
bmasem(model_syntax, sample_cov = boer16$data, sample_nobs = boer16$n)
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
