% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cooke16}
\alias{cooke16}
\title{Another dataset from the metaSEM package.}
\format{
\subsection{\code{cooke16}}{

A list of data with the following structure:
\describe{
\item{data}{
A list of correlation matrices.
The variables are SN (subjective norm), ATT (attitude),
PBC (perceived behavior control), BI (behavioral intention),
and BEH (behavior).
}
\item{n}{A vector of sample sizes}
\item{MeanAge}{
Mean age of the participants except for Ajzen and Sheikh (2013),
which is the median age, and Glassman, et al. (2010a) to
Glassman, et al. (2010d), which are based on the range of 18 to 24.
}
\item{Female}{Percentage of female participants.}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/metaSEM/index.html}
}
\usage{
cooke16
}
\description{
The data set includes 34 correlation matrices on the
how well the theory of planned behaviour predicts alcohol consumption
\insertCite{cooke_how_2016}{bayesianmasem}.
We adjusted the third correlation matrix to the nearest
valid correlation matrix.
Data and documentation copied over from the metaSEM package
\insertCite{Cheung-metaSEM}{bayesianmasem}.
}
\examples{
\dontrun{
model_syntax <- paste(
  "BI ~ ATT + SN + PBC",
  "BEH ~ PBC + BI",
  sep = "\n"
)
bmasem(model_syntax, sample_cov = cooke16$data, sample_nobs = cooke16$n)
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
