% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bayesianmasem_version}
\alias{.bayesianmasem_version}
\title{Get package version function}
\usage{
.bayesianmasem_version()
}
\value{
Package version as string
}
\description{
Get package version function
}
\keyword{internal}
