% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bmasem_post_sum_draws}
\alias{.bmasem_post_sum_draws}
\title{Posterior summary worker function}
\usage{
.bmasem_post_sum_draws(draws, interval = 0.9)
}
\arguments{
\item{draws}{Posterior draws}

\item{interval}{Confidence interval to select}
}
\value{
Summary of posterior draws
}
\description{
A worker that slightly modifies the default summary function
in posterior package.
}
\keyword{internal}
