% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_bm_data}
\alias{.check_bm_data}
\title{mini-check for data inputs}
\usage{
.check_bm_data(data, group, sample_cov, sample_nobs)
}
\arguments{
\item{sample_cov}{(list of matrices) sample covariance or correlation
matrices.
The rownames and/or colnames must contain the observed variable names.}

\item{sample_nobs}{(vector of positive integer) Number of observations
for each study.}
}
\description{
mini-check for data inputs
}
\keyword{internal}
