% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_bm_method}
\alias{.check_bm_method}
\title{mini-check for estimation method}
\usage{
.check_bm_method(method)
}
\arguments{
\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", or "none". See details below.}
}
\description{
mini-check for estimation method
}
\keyword{internal}
