% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_bm_type}
\alias{.check_bm_type}
\title{mini-check for type}
\usage{
.check_bm_type(type)
}
\arguments{
\item{type}{(character) One of "fe", "re", or "dep" for fixed-effects,
random-effects, and dependent-samples MASEM respectively.
The "dep" argument is experimental, see details below.}
}
\description{
mini-check for type
}
\keyword{internal}
