% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_list_pa_meta.R
\name{.create_data_list_pa_meta}
\alias{.create_data_list_pa_meta}
\title{Stan data helper function}
\usage{
.create_data_list_pa_meta(
  lavaan_object = NULL,
  method = "normal",
  type = "re",
  priors = NULL,
  cluster = NULL,
  partab = NULL,
  x_mat = NULL,
  conditional_re = TRUE,
  old_data = NULL
)
}
\arguments{
\item{lavaan_object}{lavaan fit object of corresponding model}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", or "none". See details below.}

\item{type}{(character) One of "fe", "re", or "dep" for fixed-effects,
random-effects, and dependent-samples MASEM respectively.
The "dep" argument is experimental, see details below.}

\item{priors}{An object of \code{\link{bmasempriors-class}}.
See \code{\link{new_bmasempriors}} for more information.}

\item{cluster}{An optional integer vector identifying the cluster each group
belongs to.
Asssume there are five groups, the first three belong to cluster 1
and the last two belong to cluster 2,
then the argument would be: \code{cluster = c(1, 1, 1, 2, 2)}.
This feature is experimental, see details below.}

\item{partab}{lavaan parameter table output with ceq.simple & std.lv = TRUE}

\item{x_mat}{(data.frame)
Meta-analytic predictors. Each row contains the data for a group and
each column is a variable. Column names should be labelled.}

\item{conditional_re}{(LOGICAL)
Only relevant for analysis of correlation structures.
If TRUE, sample levels of the study-level random effect (usually faster);
If FALSE, don't (usually more efficient).}

\item{old_data}{original list of matrices, needed for PA}
}
\value{
Data list object used in fitting Stan model
}
\description{
A function that creates data list object passed to Stan
}
\keyword{internal}
