% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_list_pa_pooled.R
\name{.create_data_list_pa_pooled}
\alias{.create_data_list_pa_pooled}
\title{Stan data helper function}
\usage{
.create_data_list_pa_pooled(
  lavaan_object = NULL,
  method = "normal",
  priors = NULL,
  partab = NULL,
  acov_mat = NULL,
  old_names = NULL
)
}
\arguments{
\item{lavaan_object}{lavaan fit object of corresponding model}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", or "none". See details below.}

\item{priors}{An object of \code{\link{bmasempriors-class}}.
See \code{\link{new_bmasempriors}} for more information.}

\item{partab}{lavaan parameter table output with ceq.simple & std.lv = TRUE}

\item{acov_mat}{(Optional) The asymptotic variance matrix of
lower triangular half (column-order) of the correlation matrix
to be used for correlation structure analysis.}

\item{old_names}{(Optional) Variable name order of original
correlation matrix, used to reorder acov_mat.}
}
\value{
Data list object used in fitting Stan model
}
\description{
Stan data helper function
}
\keyword{internal}
