% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_list_pooled.R
\name{.create_data_list_pooled}
\alias{.create_data_list_pooled}
\title{Stan data helper function}
\usage{
.create_data_list_pooled(
  lavaan_object = NULL,
  method = "normal",
  simple_struc = TRUE,
  priors = NULL,
  partab = NULL,
  acov_mat = NULL
)
}
\arguments{
\item{lavaan_object}{lavaan fit object of corresponding model}

\item{method}{(character) One of "normal", "lasso", "logistic",
"GDP", or "none". See details below.}

\item{simple_struc}{(LOGICAL) Only relevant for CFAs.
If TRUE (default): assume simple structure;
If FALSE: estimate all cross-loadings using generalized}

\item{priors}{An object of \code{\link{bmasempriors-class}}.
See \code{\link{new_bmasempriors}} for more information.}

\item{partab}{lavaan parameter table output with ceq.simple & std.lv = TRUE}

\item{acov_mat}{(Optional) The asymptotic variance matrix of
lower triangular half (column-order) of the correlation matrix
to be used for correlation structure analysis.}
}
\value{
Data list object used in fitting Stan model
}
\description{
A function that creates data list object passed to Stan
}
\keyword{internal}
