% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.frechet_log_apply}
\alias{.frechet_log_apply}
\title{Frechet derivative of matrix logarithm}
\usage{
.frechet_log_apply(h_mat, prep)
}
\arguments{
\item{h_mat}{Matrix to apply derivative}

\item{prep}{List from \code{\link{.frechet_log_prep}}}
}
\value{
Matrix of same dimensions as h_mat
}
\description{
Frechet derivative of matrix logarithm
}
\keyword{internal}
