% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.frechet_log_prep}
\alias{.frechet_log_prep}
\title{Prepare for Frechet derivative of matrix logarithm}
\usage{
.frechet_log_prep(r_mat)
}
\arguments{
\item{r_mat}{(matrix) Correlation matrix}
}
\value{
A list with:
\item{q_vecs}{Eigenvectors of \code{r_mat}}
\item{f_mat}{Matrix for Frechet derivative computation}
}
\description{
Prepare for Frechet derivative of matrix logarithm
}
\keyword{internal}
