% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_log_mat_missing}
\alias{.get_log_mat_missing}
\title{A function to impute a correlation matrix with missing elements}
\usage{
.get_log_mat_missing(s_mat, n = 1, correlation = TRUE, n_imps = 2000)
}
\arguments{
\item{s_mat}{(matrix) Correlation or covariance matrix}

\item{n}{Sample size}

\item{correlation}{(LOGICAL) If TRUE, assume correlation matrix;
If FALSE, assume covariance matrix.}
}
\value{
Returns a list with the lower half vector of log(correlation matrix)
and its asymptotic variance matrix
}
\description{
A function to impute a correlation matrix with missing elements
}
\keyword{internal}
