% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.init_fx}
\alias{.init_fx}
\title{Initial values for Stan fitter}
\usage{
.init_fx(dl)
}
\arguments{
\item{data_list}{A data list object}
}
\value{
A function with initial values
}
\description{
A function that provides initial values for Stan fitter
}
\keyword{internal}
