% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ivecl}
\alias{.ivecl}
\title{Off-diagonal to full transformation}
\usage{
.ivecl(v)
}
\arguments{
\item{v}{(vector) Off-diagonal elements}
}
\value{
Full matrix
}
\description{
Off-diagonal to full transformation
}
\keyword{internal}
