% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.process_x_mat}
\alias{.process_x_mat}
\title{A function to process the meta-analytic SEM predictor matrix}
\usage{
.process_x_mat(x_mat = NULL, n_groups)
}
\arguments{
\item{x_mat}{(data.frame)
Meta-analytic predictors. Each row contains the data for a group and
each column is a variable. Column names should be labelled.}

\item{n_groups}{Number of groups}
}
\value{
Returns numeric matrix of predictor variables
}
\description{
A function to process the meta-analytic SEM predictor matrix
}
\keyword{internal}
