% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g_inv_map}
\alias{g_inv_map}
\title{Transform unbounded vector to correlation matrix}
\usage{
g_inv_map(gamma_in, tol_val = 1e-13, ret_vec = FALSE)
}
\arguments{
\item{gamma_in}{(vector) Strict lower half vector of log(correlation matrix)}

\item{tol_val}{(positive real) Tolerance for calculation}

\item{ret_vec}{(Logical) If TRUE, return strict lower half vector,
ELSE: return correlation matrix.}
}
\value{
Transformed correlation matrix
}
\description{
Transform unbounded vector to correlation matrix
}
