% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_defs.R
\name{new_bmasempriors}
\alias{new_bmasempriors}
\title{Set priors in package}
\usage{
new_bmasempriors(
  lkj_shape = 2,
  ml_par = 0,
  sl_par = 0.5,
  rs_par = 1,
  rc_par = 2,
  mr_par = log(0.8),
  sr_par = 0.7,
  br_par = 0.5,
  rm_par = 0.15
)
}
\arguments{
\item{lkj_shape}{(positive real) The shape parameter of the LKJ-prior on the
interfactor correlation matrix in confirmatory factor models.}

\item{ml_par}{(real) The location parameter of the normal prior on loadings.}

\item{sl_par}{(positive real) The scale parameter
of the normal prior on loadings.}

\item{rs_par}{(positive real) The scale parameter
of the Student-t(3,0,) prior on residual standard deviations.}

\item{rc_par}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
prior on the residual error correlations.}

\item{mr_par}{(real) The location parameter of the normal prior
on the log-RMSEA.}

\item{sr_par}{(positive real) The scale parameter of the normal prior
on the log-RMSEA.}

\item{br_par}{(positive real) The scale parameter of the normal prior
on the regression coefficients for the log-RMSEA.}

\item{rm_par}{(positive real) The scale parameter of the normal prior
on the tau / CRMR parameter.}
}
\value{
An object of \code{\link{bmasempriors-class}}
}
\description{
Modify default priors in package.
}
\examples{
\dontrun{
# Change LKJ shape parameter only
custom_priors <- new_bmasempriors(lkj_shape = 1.0)
model_syntax <- paste0(
  "distress =~ ", paste0("x", 1:14, collapse = " + "), "\n",
  "anxiety =~ ", paste0("x", seq(1, 14, 2), collapse = " + "), "\n",
  "depression =~ ", paste0("x", seq(2, 14, 2), collapse = " + ")
)
bmasem(
  model_syntax,
  sample_cov = Norton13$data, sample_nobs = Norton13$n, orthogonal = TRUE,
  priors = custom_priors
)
}
}
