% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_summary.R
\name{pp_summary}
\alias{pp_summary}
\title{Pretty print model results}
\usage{
pp_summary(object, interval = 0.9, digits = 3, simple = TRUE, save_html = NULL)
}
\arguments{
\item{object}{(bmasem) An object of \code{\link{bmasem-class}}
returned by \code{\link{bmasem}}.}

\item{interval}{(real in (0, 1)) Credible interval to return.}

\item{digits}{(positive integer) Number of decimal places to print in table}

\item{simple}{(Logical) TRUE to produce table with less information
about parameters;
FALSE: produces table with more information}

\item{save_html}{(string) Optional file name to save table as HTML}
}
\description{
Nice printing of model results,
optionally produces HTML document
}
\examples{
\dontrun{
model_syntax <- paste0(
  "distress =~ ", paste0("x", 1:14, collapse = " + "), "\n",
  "anxiety =~ ", paste0("x", seq(1, 14, 2), collapse = " + "), "\n",
  "depression =~ ", paste0("x", seq(2, 14, 2), collapse = " + ")
)
bmasem(
  model_syntax,
  sample_cov = Norton13$data, sample_nobs = Norton13$n, orthogonal = TRUE
)
pp_summary(fit)
pp_summary(fit, simple = FALSE)
}
}
