% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_bm_data}
\alias{.check_bm_data}
\title{mini-check for data inputs}
\usage{
.check_bm_data(data, group, sample_cov, sample_nobs)
}
\arguments{
\item{data}{An optional data frame containing the observed variables used in
the model.}

\item{group}{An optional string identifying the grouping variable in
the data object.}

\item{sample_cov}{(list of matrices) sample covariance or correlation
matrices.
The rownames and/or colnames must contain the observed variable names.
For now, assumes there are no missing elements in the covariance matrices.}

\item{sample_nobs}{(vector of positive integer) Number of observations
for each study.}
}
\description{
mini-check for data inputs
}
\keyword{internal}
